API/requirements.txt
fastapi
uvicorn
jinja2
httpx
fastf1
pycountry
pytz
fastapi-cache2
svgwrite
numpy
